Got rid of shared_ptr Fixed memory leaks Modified some game scripts Updated compilation scripts

+++ src/include/UIElement.h
@@ -33,7 +33,7 @@ public:
 
 protected:
 	int x, y;
-	int width, height;
+	size_t width, height;
 	Level* level;
 };
